% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{All_Flight_Left}
\alias{All_Flight_Left}
\title{Drone flight Left}
\format{
A data frame with 1676 rows and 17 variables:
\describe{
\item{\code{Photo No.}}{character COLUMN_DESCRIPTION}
\item{\code{Date}}{double COLUMN_DESCRIPTION}
\item{\code{Time}}{double COLUMN_DESCRIPTION}
\item{\code{Lat_dec}}{double COLUMN_DESCRIPTION}
\item{\code{Lon_dec}}{double COLUMN_DESCRIPTION}
\item{\code{Alt above the Sea level}}{double COLUMN_DESCRIPTION}
\item{\code{Alt above launch point}}{double COLUMN_DESCRIPTION}
\item{\code{Alt above ground}}{double COLUMN_DESCRIPTION}
\item{\code{Banking angle}}{character COLUMN_DESCRIPTION}
\item{\code{Tangage}}{character COLUMN_DESCRIPTION}
\item{\code{Azimuth}}{character COLUMN_DESCRIPTION}
\item{\code{GPS Course}}{character COLUMN_DESCRIPTION}
\item{\code{Speed}}{character COLUMN_DESCRIPTION}
\item{\code{GPS Speed}}{character COLUMN_DESCRIPTION}
\item{\code{Camera position}}{character COLUMN_DESCRIPTION}
\item{\code{Species}}{logical COLUMN_DESCRIPTION}
\item{\code{Number}}{logical COLUMN_DESCRIPTION}
}
}
\usage{
All_Flight_Left
}
\description{
Dataset of drone flight
}
\details{
DETAILS
}
\keyword{datasets}
