% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{All_Flight_Right}
\alias{All_Flight_Right}
\title{DATASET_TITLE}
\format{
A data frame with 12670 rows and 18 variables:
\describe{
\item{\code{id}}{character COLUMN_DESCRIPTION}
\item{\code{Photo No.}}{character COLUMN_DESCRIPTION}
\item{\code{Date}}{character COLUMN_DESCRIPTION}
\item{\code{Time}}{double COLUMN_DESCRIPTION}
\item{\code{Lat_dec}}{double COLUMN_DESCRIPTION}
\item{\code{Lon_dec}}{double COLUMN_DESCRIPTION}
\item{\code{Alt above the Sea level}}{double COLUMN_DESCRIPTION}
\item{\code{Alt above launch point}}{double COLUMN_DESCRIPTION}
\item{\code{Alt above ground}}{double COLUMN_DESCRIPTION}
\item{\code{Banking angle}}{double COLUMN_DESCRIPTION}
\item{\code{Tangage}}{double COLUMN_DESCRIPTION}
\item{\code{Azimuth}}{double COLUMN_DESCRIPTION}
\item{\code{GPS Course}}{double COLUMN_DESCRIPTION}
\item{\code{Speed}}{double COLUMN_DESCRIPTION}
\item{\code{GPS Speed}}{double COLUMN_DESCRIPTION}
\item{\code{Camera position}}{character COLUMN_DESCRIPTION}
\item{\code{Species}}{logical COLUMN_DESCRIPTION}
\item{\code{Number}}{logical COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
All_Flight_Right
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
