% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alt_threshold.R
\name{Alt_threshold}
\alias{Alt_threshold}
\title{Apply a threshold to the lower locations that occur at takeoff and landing}
\usage{
Alt_threshold(df = df, Alt = Alt, height = 100)
}
\arguments{
\item{Alt}{Altitude column in the dataframe}

\item{height}{The threshold height (remove values lower than the height)}
}
\description{
Apply a threshold to the lower locations that occur at takeoff and landing
}
