% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_DEM_values.R
\docType{data}
\name{Flight1_LeftCamera}
\alias{Flight1_LeftCamera}
\title{Get raster values from DEM}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1676 rows and 17 columns.
}
\usage{
Flight1_LeftCamera
}
\arguments{
\item{Flight_file}{The flight file as a dataframe}

\item{Image_Width}{Width of image}

\item{Focal_Length}{Focal length}

\item{Angle}{Angle of the camera}
}
\value{
Ground surface distance
}
\description{
Get raster values from DEM
}
\keyword{datasets}
