% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_overlap.R
\name{forward_overlap}
\alias{forward_overlap}
\title{Calculate forward overlap}
\usage{
forward_overlap(
  altitude1,
  altitude2,
  forward_distance,
  angle_of_camera = 25,
  banking_angle1,
  banking_angle2
)
}
\arguments{
\item{altitude1}{Drone altitude at position1}

\item{altitude2}{Drone altitude at position2}

\item{forward_distance}{distance between points where image was taken}

\item{angle_of_camera}{angle of the camera (25 or 25.7 degrees)}

\item{banking_angle1}{The banking angle in the first position}

\item{banking_angle2}{The banking angle in the second position}
}
\description{
Calculate forward overlap
}
