% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsd.R
\name{get_GSD}
\alias{get_GSD}
\title{Calculate the ground surface distance}
\usage{
get_GSD(
  Sensor_Height = 24,
  Sensor_Width = 35,
  Image_Width = 7952,
  Image_Height = 5304,
  Focal_Length = 35,
  Altitude
)
}
\arguments{
\item{Sensor_Height}{Height of sensor}

\item{Sensor_Width}{Width of sensor}

\item{Image_Width}{Width of image}

\item{Image_Height}{Height of image}

\item{Focal_Length}{Focal length}

\item{altitude}{the height of the drone above the ground}
}
\value{
Ground surface distance
}
\description{
Calculate the ground surface distance
}
