% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field_of_view.R
\name{get_HFOV}
\alias{get_HFOV}
\title{Get Horizontal Field of view}
\usage{
get_HFOV(HSensor = 35.9, Focallen = 35)
}
\arguments{
\item{HSensor}{Horizontal sensor in mm}

\item{Focallen}{Focal length of the camera}
}
\description{
Get Horizontal Field of view
}
