% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo_area.R
\name{get_photo_area}
\alias{get_photo_area}
\title{Calculate the area of a drone image}
\usage{
get_photo_area(altitude, angle_of_camera = 25, banking_angle)
}
\arguments{
\item{altitude}{= Drone altitude}

\item{angle_of_camera}{The set angle of the camera}
}
\value{
The area of the photograph in m2
}
\description{
Calculate the area of a drone image
}
