% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strip_width.R
\name{get_strip_width}
\alias{get_strip_width}
\title{Calculate the strip width (two cameras on the drone)}
\usage{
get_strip_width(
  altitude,
  angle1 = 25,
  angle2 = 25.7,
  distance_between_cameras = 0.03
)
}
\arguments{
\item{altitude}{the height of the drone above the ground}

\item{angle1}{the angle of camera1}

\item{angle2}{the angle of camera2}

\item{distance_between_cameras}{added value for the distance between cameras}
}
\value{
distance along the ground
}
\description{
Calculate the strip width (two cameras on the drone)
}
