% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field_of_view.R
\name{get_VFOV}
\alias{get_VFOV}
\title{Get Vertical Field of view}
\usage{
get_VFOV(VSensor = 24, Focallen = 35)
}
\arguments{
\item{VSensor}{Vertical sensor in mm}

\item{Focallen}{Focal length of the camera}
}
\description{
Get Vertical Field of view
}
