% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd_in_image.R
\name{grd_in_img}
\alias{grd_in_img}
\title{Calculate the ground in the image (two cameras on the drone)}
\usage{
grd_in_img(
  altitude,
  angle1 = 25,
  angle2 = 25.7,
  distance_between_cameras = 0.03
)
}
\arguments{
\item{altitude}{the height of the drone above the ground}

\item{angle1}{the angle of camera1}

\item{angle2}{the angle of camera2}

\item{distance_between_cameras}{added value for the distance between cameras}
}
\value{
distance along the ground
}
\description{
Calculate the ground in the image (two cameras on the drone)
}
