% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_resolution.R
\name{Ground_footprint}
\alias{Ground_footprint}
\title{Calculate the ground footprint}
\usage{
Ground_footprint(GSD = GSD, Image_Height = 7952, Image_Width = 5304)
}
\arguments{
\item{GSD}{Spatial resolution}

\item{Image_Height}{Height of image}

\item{Image_Width}{Width of image}
}
\value{
Ground footprint area (no angular distortion)
}
\description{
Calculate the ground footprint
}
\details{
This function calculates the ground footprint with no angular distortion. The width and the length of the sensor are separately multiplied by the ground sampling distance.
Then the resulting width and length are multiplied to return the area of the image
}
