% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo_area.R
\name{ground_in_image}
\alias{ground_in_image}
\title{Calculate the area of a drone image}
\usage{
ground_in_image(alt, AOV)
}
\arguments{
\item{alt}{= Drone altitude}

\item{AOV}{= Angle of view}
}
\description{
Calculate the area of a drone image
}
