% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_resolution.R
\name{GSD}
\alias{GSD}
\title{Ground sampling distance (Spatial Resolution)}
\usage{
GSD(f = 35, H = Altitude, pixel = 4.5)
}
\arguments{
\item{H}{the height of the drone above the ground}

\item{pixel}{pixel size in micrometers}

\item{Focal_Length}{Focal length}
}
\value{
Ground surface distance
}
\description{
Calculates the ground sampling distance (spatial resolution). This version of the function ignores angular distortion
}
\details{
This version of the function assumes no angular distortion. The spatial resolution is calculated using the scale (focal length of the camera / Flying height above ground level). This equates to the sensor pixel size/ Ground pixel size. To calculate the ground sample distance we need to know the flying height, the camera sensor size and the lens focal length.

Camera: SONY DCS-RX1RM2

Sensor: 35.9 x 24.0 mm

Viewing angle: (CORRESPONDING 35 MM FORMAT) 63 degrees (35mm))

Focal length: 35mm

Flight height: 200-250m

Pixels: 4.5 micrometers

Image width in pixels: 7952

Image height in pixels: 5304
}
