% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo_area.R
\name{Image_sides}
\alias{Image_sides}
\title{Calculate the length of the sides of image}
\usage{
Image_sides(Altitude = Altitude)
}
\arguments{
\item{alt}{= Drone altitude}
}
\description{
Calculate the length of the sides of image
}
