% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolly2.R
\name{jolly2}
\alias{jolly2}
\title{Calculate Jolly II estimate and 95\% Confidence Interval}
\usage{
jolly2(species_count, Transect_area, Z = Z)
}
\arguments{
\item{species_count}{the column from the dataset with the species counts in}

\item{Transect_area}{the column from the dataset with the transect areas in}

\item{Z}{the stratum area (the total area that we are extrapolating to)}
}
\value{
dataframe with Jolly II estimate and 95\% CI
}
\description{
Calculate Jolly II estimate and 95\% Confidence Interval
}
\examples{
data=Kulan::sim_kulan_data(n=100,ma=25, p0=0.5, ml=20, w=0.2)
jolly2(data$sp_count, data$Trans_area, Z=164*5)
}
