% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_overlap.R
\name{oforward}
\alias{oforward}
\title{Calculate forward overlap}
\usage{
oforward(dforward, f = 35, H = H, w = 35.9)
}
\arguments{
\item{H.}{= Drone altitude}

\item{dforward.}{= distance between points where image was taken}

\item{f.=}{focal length (default = 35mm)}

\item{w.=}{sensor width (default = 24mm)}
}
\description{
Calculate forward overlap
}
\references{
\url{https://doi.org/10.3390/rs11101252}
}
