% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_overlap.R
\name{oside}
\alias{oside}
\title{Calculate side overlap}
\usage{
oside(dside, f = 35, H = H, w = 24)
}
\arguments{
\item{H.}{= Drone altitude}

\item{dside.}{= distance between points where image was taken}

\item{f.=}{focal length (default = 35mm)}

\item{w.=}{sensor width (default = 24mm)}
}
\description{
Calculate side overlap
}
\references{
\url{https://doi.org/10.3390/rs11101252}
}
