% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo_area.R
\name{photo_area}
\alias{photo_area}
\title{Calculate the area of a drone image}
\usage{
photo_area(alt, AOV)
}
\arguments{
\item{alt}{= Drone altitude}

\item{AOV}{= Angle of view}
}
\description{
Calculate the area of a drone image
}
