% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trapezoid.R
\name{plot_trapezoid}
\alias{plot_trapezoid}
\title{Plot the dimensions of the photograph}
\usage{
plot_trapezoid(altitude, angle_of_camera = 25, banking_angle)
}
\arguments{
\item{altitude}{= Drone altitude}

\item{angle_of_camera}{The set angle of the camera}

\item{banking_angle}{The banking angle of the drone}
}
\value{
plot of the horizontal and vertical profile of the photograph and the
resulting trapezoid
}
\description{
Plot the dimensions of the photograph
}
