% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kulan_model.R
\name{run_kulan_model}
\alias{run_kulan_model}
\title{Calculate Bayesian estimate and 95\% HDI}
\usage{
run_kulan_model(
  p_area = p_area.,
  p_min = p_min.,
  p_max = p_max.,
  no_ind = no_ind.,
  plot.out = "No"
)
}
\arguments{
\item{p_area}{proportion of the stratum that has been sampled}

\item{p_min}{prior minimum number of animals}

\item{p_max}{prior maximum number of animals}

\item{no_ind}{Indivduals count column from data}

\item{plot.out}{Add a plot ("Yes") of the distribution of the estimate (default is "No")}
}
\value{
dataframe with estimate and hdi
}
\description{
Calculate Bayesian estimate and 95\% HDI
}
