% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_overlap.R
\name{side_overlap}
\alias{side_overlap}
\title{Calculate the side overlap}
\usage{
side_overlap(
  altitude,
  angle_of_camera1 = 25,
  angle_of_camera2 = 25.7,
  banking_angle
)
}
\arguments{
\item{altitude}{= Drone altitude}

\item{angle_of_camera1}{The set angle of the camera on the left
#' @param angle_of_camera2 The set angle of the camera on the right}

\item{banking_angle}{The banking angle of the drone}
}
\value{
side overlap in meters
}
\description{
Calculate the side overlap
}
\examples{
side_overlap(altitude = 200, banking_angle = 0)
}
