% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_kulan_data.R
\name{sim_kulan_data}
\alias{sim_kulan_data}
\title{Simulate a dataset}
\usage{
sim_kulan_data(n = 100, ma = 25, p0 = 0.5, ml = 20, w = 0.2)
}
\arguments{
\item{n}{number of transects}

\item{ma}{mean number of animals counted where they are counted}

\item{p0}{proportion of the transects that will be empty}

\item{ml}{mean length of the transects}

\item{w}{width of the transects (assuming a constant width)}
}
\value{
simulated data in a dataframe
}
\description{
Simulate a dataset
}
\examples{
sim_kulan_data(n=100,ma=25, p0=0.5, ml=20, w=0.2)
}
