% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_dark_blue}
\alias{theme_dark_blue}
\title{Theme choices: theme_publication https://github.com/koundy/ggplot_theme_Publication}
\usage{
theme_dark_blue(base_size = 14, base_family = "sans")
}
\description{
Theme choices: theme_publication https://github.com/koundy/ggplot_theme_Publication
}
