% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strip_width.R
\name{get_strip_width}
\alias{get_strip_width}
\title{Calculate the strip width (two cameras on the drone)}
\usage{
get_strip_width(
  xsensor = 35.9,
  ysensor = 24,
  focallen = 35,
  ygim = 0,
  xgim = 25,
  alt,
  banking_angle
)
}
\arguments{
\item{xsensor}{horizontal sensor in mm}

\item{ysensor}{vertical sensor in mm}

\item{focallen}{focal length of the camera}

\item{ygim}{angle of the camera on the y axis}

\item{xgim}{angle of the camera on the x axis}

\item{alt}{the altitude of the drone}
}
\value{
value of the horizontal strip-width (one side)
}
\description{
Calculate the strip width (two cameras on the drone)
}
