% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tables}
\alias{get_tables}
\title{Extract all csv files from a nested folder on your local computer}
\usage{
get_tables(path = path)
}
\arguments{
\item{path}{the path to the nested folders}
}
\description{
Extract all csv files from a nested folder on your local computer
}
