% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_resolution.R
\name{GSD}
\alias{GSD}
\title{Ground sampling distance (Spatial Resolution)}
\usage{
GSD(altitude, angle_of_camera = 25, pixel = 4.5, focallen = 35)
}
\arguments{
\item{altitude}{the height of the drone above the ground}

\item{angle_of_camera}{the angle of the camera is set}

\item{pixel}{pixel size in micrometers}

\item{focallen}{Focal length}
}
\value{
Ground surface distance m per pixel
}
\description{
Calculates the ground sampling distance (spatial resolution). This version of the function ignores angular distortion
}
\details{
Camera: SONY DCS-RX1RM2

Sensor: 35.9 x 24.0 mm

Viewing angle: (CORRESPONDING 35 MM FORMAT) 63 degrees (35mm))

Focal length: 35mm

Flight height: 200-250m

Pixels: 4.5 micrometers

Image width in pixels: 7952

Image height in pixels: 5304
}
