% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_overlap.R
\name{oforward}
\alias{oforward}
\title{Calculate forward overlap}
\usage{
oforward(dforward = dforward, f = 35, H = H, w = 24)
}
\arguments{
\item{dforward}{= distance between points where image was taken}

\item{H}{= Drone altitude}

\item{f=}{focal length (35mm)}

\item{w=}{sensor width (24mm)}
}
\description{
Calculate forward overlap
}
\references{
\url{https://doi.org/10.3390/rs11101252}
}
