% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo_area.R
\name{photo_area}
\alias{photo_area}
\title{Calculate the area of a drone image}
\usage{
photo_area(X, theta, r)
}
\arguments{
\item{theta}{= Field of view}

\item{x}{= Drone height}
}
\description{
Calculate the area of a drone image
}
\details{
If the diagonal of the image is D, D/2 is the length of the base
of the right-angled triangle with the two included angles as θ°/2
a (180°-θ°)/2 (as becomes clear when the FoV angle is bisected
to create two identical right-angled triangles).
f (A) and (B) are the sides of the image, then the aspect ratio
(r) is equal to either A/B or B/A. We assume (A) to be the
independent variable and (B) to be the dependent variable.
}
